refactor(harness): centralize policy approval and package security
